Added separate logback configuration file for tests
There is now a separate `logback-test.xml` configuration file for tests.
It outputs log in console in more readable format than the JSON format
used in production code. This is helpful when log output is needed during
tests, for example to find out why test is failing. By default, logging
is set off to not clutter the screen with unnecessary noise.