Added handling for ResourceNotFoundException in deleteTable
Previously, function `deleteTable` would throw ResourceNotFoundException
when table `AuditLog` did not exist in database (as is the case when
setting up local environment for the first time, for example). Now that
exception is caught and a log entry is emitted instead. This way the
function is safe to use whether or not the table exists.